refactor: simplify shared document access and validation
Closes #398
