+++ entrypoints/popup/queries/__tests__/settings.test.tsx
@@ -7,7 +7,8 @@ import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { fakeBrowser } from 'wxt/testing/fake-browser';
 import { sendMessage } from '@/infrastructure/browser/messages';
 import { readLearningDocument, replaceLearningDocument } from '@/infrastructure/storage/learning-document';
-import { getSettings, updateSettings } from '@/services/settings';
+import { getSettings } from '@/infrastructure/storage/learning-queries';
+import { updateSettings } from '@/services/settings';
 import { createMessageMock } from '@/test/utils/message-mocks';
 import { createTestWrapper } from '@/test/utils/test-wrapper';
 import { cardQueryKeys } from '../cards';
